docs: update readme.md
