+++ src/components/Modal/index.ts
@@ -1,2 +1,3 @@
 export * from "./Modal";
 export * from "./ModalContext";
+export * from "./ModalProvider";
